optimisation of gotenbergService file
