+++ module/userProjects/helper.js
@@ -2716,6 +2716,80 @@ module.exports = class UserProjectsHelper {
 	static createCertificate(certificateData, projectId, asyncMode = true) {
 		return new Promise(async (resolve, reject) => {
 			try {
+				// Define the temporary folder path for the certificate
+				const certificateTempFolderPath = path.resolve(__dirname + '/../../certificateTempFolder')
+
+				// Remove the temporary folder if it exists
+				if (fs.existsSync(certificateTempFolderPath)) {
+					fs.rmSync(certificateTempFolderPath, { recursive: true, force: true })
+				}
+
+				// Create the temporary folder
+				fs.mkdirSync(certificateTempFolderPath)
+
+				// Generate UUID for svg file
+				const dynamicUUIDForSvgFile = UTILS.generateUniqueId()
+
+				// Define the path for the SVG template
+				const svgTemplatePath = certificateTempFolderPath + `/${dynamicUUIDForSvgFile}_template.svg`
+
+				// Write the populated SVG template to the file
+				fs.writeFileSync(svgTemplatePath, certificateData.populatedSVGTemplate)
+
+				// Read the SVG template and convert it to a single line string
+				fs.readFileSync(svgTemplatePath, 'utf-8', (error, data) => {
+					if (error) {
+						// If there's an error reading the file, clean up and throw an error
+						if (fs.existsSync(certificateTempFolderPath)) {
+							fs.rmSync(certificateTempFolderPath, { recursive: true, force: true })
+						}
+						throw {
+							message: CONSTANTS.apiResponses.CERTIFICATE_GENERATION_FAILED,
+						}
+					}
+
+					// Replace all whitespace with a single space and trim the string
+					const singleLineSvg = data.replace(/\s+/g, ' ').trim()
+
+					// Write the single line SVG back to the file
+					fs.writeFileSync(svgTemplatePath, singleLineSvg, 'utf-8', (error) => {
+						if (error) {
+							// If there's an error writing the file, clean up and throw an error
+							if (fs.existsSync(certificateTempFolderPath)) {
+								fs.rmSync(certificateTempFolderPath, { recursive: true, force: true })
+							}
+							throw {
+								message: CONSTANTS.apiResponses.CERTIFICATE_GENERATION_FAILED,
+							}
+						}
+					})
+				})
+
+				const svgUploadDetails = await common_handler.uploadPdfToCloud(
+					`${dynamicUUIDForSvgFile}_template.svg`,
+					certificateData.userId,
+					certificateTempFolderPath
+				)
+
+				if (!svgUploadDetails || svgUploadDetails.data == '') {
+					throw {
+						message: CONSTANTS.apiResponses.FAILED_TO_CREATE_DOWNLOADABLEURL,
+					}
+				}
+
+				// Update the project in the database
+				await projectQueries.findOneAndUpdate(
+					{
+						_id: projectId,
+					},
+					{
+						$set: {
+							'certificate.svgPath': svgUploadDetails.data,
+						},
+					}
+				)
+				certificateData['svgTemplatePath'] = svgTemplatePath
+				certificateData['certificateTempFolderPath'] = certificateTempFolderPath
 				// Call the service to create the certificate and get the details
 				const certificateDetails = await gotenbergService.createCertificate(certificateData, asyncMode)
 				// Check if the certificate creation was successful
@@ -2736,9 +2810,38 @@ module.exports = class UserProjectsHelper {
 					updateObject['$set']['certificate.transactionIdCreatedAt'] = new Date()
 				}
 
+				if (!asyncMode) {
+					// Generate UUID for pdf file
+					const dynamicUUIDForPdfFile = UTILS.generateUniqueId()
+
+					// Define the path for the output PDF
+					const certificatePdfPath = certificateTempFolderPath + `/${dynamicUUIDForPdfFile}_output.pdf`
+
+					// Write the response data (PDF) to the file
+					fs.writeFileSync(certificatePdfPath, certificateDetails.data.gotenbergResponse.data)
+
+					// Upload the PDF and SVG template to the cloud
+					const pdfUploadDetails = await common_handler.uploadPdfToCloud(
+						`${dynamicUUIDForPdfFile}_output.pdf`,
+						certificateData.userId,
+						certificateTempFolderPath
+					)
+
+					// If the PDF upload was successful, update the project details
+					if (pdfUploadDetails.success && pdfUploadDetails.data != '') {
+						updateObject['$set']['certificate.pdfPath'] = pdfUploadDetails.data
+						updateObject['$set']['certificate.issuedOn'] = new Date()
+					}
+
+					// Clean up the temporary folder
+					if (fs.existsSync(certificateTempFolderPath)) {
+						fs.rmSync(certificateTempFolderPath, { recursive: true, force: true })
+					}
+				}
+
 				// If there are fields to update, update the project details with the certificate information
 				if (Object.keys(updateObject['$set']).length > 0) {
-					let updatedProject = await projectQueries.findOneAndUpdate(
+					await projectQueries.findOneAndUpdate(
 						{
 							_id: projectId,
 						},
